requirements.txt
numpy==1.21.5
pandas==1.3.5
pyyaml==6.0
requests==2.26.0


# Para la API de salida
pydantic==1.9.0
asgiref==3.4.1
certifi==2021.5.30
click==8.0.1
fastapi==0.70.1
h11==0.12.0
pydantic==1.9.0
starlette==0.16.0
typing-extensions
uvicorn==0.16.0